Fix formatting, package name, and pom.xml file
