+++ tensorflow/samples/pom.xml
@@ -1,48 +1,24 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
-	<parent>
-		<groupId>org.bytedeco</groupId>
-		<artifactId>javacpp-presets</artifactId>
-		<version>1.4.2-SNAPSHOT</version>
-		<relativePath>../../</relativePath>
-	</parent>
+    <groupId>org.bytedeco.javacpp-presets</groupId>
+    <artifactId>tensorflow-samples</artifactId>
+    <version>1.8.0-1.4.2-SNAPSHOT</version>
+    <name>JavaCPP Presets Samples for TensorFlow</name>
 
-	<properties>
-		<tensorflow.version>1.8.0</tensorflow.version>
-	</properties>
+    <properties>
+        <maven.compiler.target>1.7</maven.compiler.target>
+        <maven.compiler.source>1.7</maven.compiler.source>
+    </properties>
 
-	<groupId>org.bytedeco.javacpp-presets</groupId>
-	<artifactId>tensorflow-samples</artifactId>
-	<version>${tensorflow.version}-${project.parent.version}</version>
-	<name>JavaCPP Presets Samples for TensorFlow</name>
-
-	<dependencies>
-		<dependency>
-			<groupId>${project.groupId}</groupId>
-			<artifactId>tensorflow-platform</artifactId>
-			<version>${tensorflow.version}-${project.parent.version}</version>
-		</dependency>
-	</dependencies>
-
-	<build>
-		<plugins>
-			<plugin>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<executions>
-					<execution>
-						<id>default-compile</id>
-						<configuration>
-							<includes>
-								<include>org/bytedeco/javacpp/preset/tensorflow/**/*.java</include>
-							</includes>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
+    <dependencies>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>tensorflow-platform</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
 
 </project>
